media-source/SourceBuffer-appendWindowStart_removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Setting SourceBuffer#appendWindowStart after the sourceBuffer object is removed</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e){
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updateend', function(e) {
                mediaSource.removeSourceBuffer(sourceBuffer);
                t.step(function() {
                    assert_throws({name: 'InvalidStateError'},
                                  function() {
                                      sourceBuffer.appendWindowStart = 1;
                                  });
                });
                t.done();
            });
            sourceBuffer.appendBuffer(data);
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#appendWindowStart : After this object is removed from the' +
'sourceBuffers attribute of the parent media source, INVALID_STATE_ERR exception should be thrown.');
</script>
</body>
</html>
